+++ nodejs/parsers/upload.js
@@ -3,17 +3,32 @@
 // GitHub : https://github.com/Drumsticks1/TS3-LogDataViewer
 
 "use strict";
-
+// TODO: doc
 const miscFunctions = require('../miscFunctions.js');
+var Channel = require("../Channel.js");
+var Upload = require("../Upload.js");
+var checkFunctions = require("../checkFunctions.js");
+var globalVariables = require("../globalVariables.js");
 
 // Todo: Eliminate duplicates after the boundaries objects.
 module.exports = {
+
+  parseUpload: function (message, dateTime) {
+    let res = this.parseMessageUpload(message);
+
+    if (Channel.getChannelByChannelId(globalVariables.ChannelList, res.channelID) === null)
+      Channel.addChannel(globalVariables.ChannelList, res.channelID, "Unknown", "Unknown");
+
+    if (!checkFunctions.isDuplicateUpload(dateTime, res.channelID, res.filename, res.uploadedByID, res.uploadedByNickname))
+      Upload.addUpload(globalVariables.UploadList, dateTime, res.channelID, res.filename, res.uploadedByID, res.uploadedByNickname);
+  },
+
   /**
    * Parses the Upload data from the given logLine.
    * @param {string} logLine
    * @returns {{channelID: number, filename: string, uploadedByNickname: string, uploadedByID: number}} the extracted data.
    */
-  parseUpload: function (logLine) {
+  parseMessageUpload: function (logLine) {
     const boundaries = {};
 
     boundaries.channelID = [
@@ -44,12 +59,18 @@ module.exports = {
     };
   },
 
+  parseUploadDeletion: function (message) {
+    let res = this.parseMessageUploadDeletion(message);
+
+    Upload.addDeletedBy(res.channelID, res.filename, res.deletedByID, res.deletedByNickname);
+  },
+
   /**
    * Parses the UploadDeletion data from the given logLine.
    * @param {string} logLine
    * @returns {{channelID: number, filename: string, deletedByNickname: string, deletedByID: number}} extracted data.
    */
-  parseUploadDeletion: function (logLine) {
+  parseMessageUploadDeletion: function (logLine) {
     const boundaries = {};
 
     boundaries.channelID = [
@@ -76,7 +97,7 @@ module.exports = {
 
     // A longer path is logged and needs to be removed in order to get the filename only.
     // Unix
-    if (filename.includes("//") )
+    if (filename.includes("//"))
       filename = filename.substr(filename.indexOf("//") + 1);
     // Windows
     else