+++ test/nodejs/parsers/clientTest.js
@@ -8,10 +8,10 @@ const assert = require('assert');
 const clientParser = require("../../../nodejs/parsers/client.js");
 
 describe('clientParser', function () {
-  describe('parseClientConnect', function () {
+  describe('parseMessageClientConnect', function () {
     it('should return the specified object', function () {
       assert.deepEqual(
-        clientParser.parseClientConnect("client connected ''client connected Testclient'(id:0) from localhost'(id:42) from 127.0.0.1:12345"),
+        clientParser.parseMessageClientConnect("client connected ''client connected Testclient'(id:0) from localhost'(id:42) from 127.0.0.1:12345"),
         {
           clientId: 42,
           Nickname: "'client connected Testclient'(id:0) from localhost",
@@ -20,10 +20,10 @@ describe('clientParser', function () {
     });
   });
 
-  describe('parseQueryClientConnect', function () {
+  describe('parseMessageQueryClientConnect', function () {
     it('should return the specified object', function () {
       assert.deepEqual(
-        clientParser.parseQueryClientConnect("query client connected 'query client connected 'TestClient from 127.0.0.1:00001'(id:0) from 127.0.0.1:12345'(id:42)"),
+        clientParser.parseMessageQueryClientConnect("query client connected 'query client connected 'TestClient from 127.0.0.1:00001'(id:0) from 127.0.0.1:12345'(id:42)"),
         {
           clientId: 42,
           Nickname: "query client connected 'TestClient from 127.0.0.1:00001'(id:0) from 127.0.0.1",
@@ -32,11 +32,11 @@ describe('clientParser', function () {
     });
   });
 
-  describe('parseClientDisconnect', function () {
+  describe('parseMessageClientDisconnect', function () {
     describe('regular disconnect', function () {
       it('should return the specified object', function () {
         assert.deepEqual(
-          clientParser.parseClientDisconnect("client disconnected 'TestClient'(id:15) reason 'reasonmsg'"),
+          clientParser.parseMessageClientDisconnect("client disconnected 'TestClient'(id:15) reason 'reasonmsg'"),
           {
             clientId: 15,
             Nickname: 'TestClient',
@@ -51,7 +51,7 @@ describe('clientParser', function () {
     describe('ban disconnect', function () {
       it('should return the specified object', function () {
         assert.deepEqual(
-          clientParser.parseClientDisconnect("client disconnected 'TestClient'(id:15) reason 'invokerid=2 invokername=admin invokeruid=OIqF4sKapFgSlTeIZOL5U7x2Yq4= reasonmsg=It's just a test bantime=10'"),
+          clientParser.parseMessageClientDisconnect("client disconnected 'TestClient'(id:15) reason 'invokerid=2 invokername=admin invokeruid=OIqF4sKapFgSlTeIZOL5U7x2Yq4= reasonmsg=It's just a test bantime=10'"),
           {
             clientId: 15,
             Nickname: 'TestClient',
@@ -66,7 +66,7 @@ describe('clientParser', function () {
     describe('kick disconnect', function () {
       it('should return the specified object', function () {
         assert.deepEqual(
-          clientParser.parseClientDisconnect("client disconnected 'TestClient'(id:15) reason 'invokerid=2 invokername=admin invokeruid=OIqF4sKapFgSlTeIZOL5U7x2Yq4= reasonmsg=It's just a test'"),
+          clientParser.parseMessageClientDisconnect("client disconnected 'TestClient'(id:15) reason 'invokerid=2 invokername=admin invokeruid=OIqF4sKapFgSlTeIZOL5U7x2Yq4= reasonmsg=It's just a test'"),
           {
             clientId: 15,
             Nickname: 'TestClient',
@@ -79,10 +79,10 @@ describe('clientParser', function () {
     });
   });
 
-  describe('parseClientDeletion', function () {
+  describe('parseMessageClientDeletion', function () {
     it('should return the id of the deleted client (1234)', function () {
       assert.equal(
-        clientParser.parseClientDeletion("client 'abc'(id:1234) got deleted by client 'admin'(id:2)"),
+        clientParser.parseMessageClientDeletion("client 'abc'(id:1234) got deleted by client 'admin'(id:2)"),
         1234
       );
     });