+++ test/nodejs/parsers/serverGroupTest.js
@@ -10,10 +10,10 @@ const serverGroupParser = require("../../../nodejs/parsers/serverGroup.js");
 // Todo: Add additional tests
 
 describe('serverGroupParser', function () {
-  describe('parseServerGroupCreation', function () {
+  describe('parseMessageServerGroupCreation', function () {
     it('should return the specified object', function () {
       assert.deepEqual(
-        serverGroupParser.parseServerGroupCreation("servergroup 'new servergroup'(id:2) was added by 'admin'(id:2)"),
+        serverGroupParser.parseMessageServerGroupCreation("servergroup 'new servergroup'(id:2) was added by 'admin'(id:2)"),
         {
           ServerGroupID: 2,
           ServerGroupName: 'new servergroup'
@@ -21,10 +21,10 @@ describe('serverGroupParser', function () {
     });
   });
 
-  describe('parseServerGroupDeletion', function () {
+  describe('parseMessageServerGroupDeletion', function () {
     it('should return the specified object', function () {
       assert.deepEqual(
-        serverGroupParser.parseServerGroupDeletion("servergroup 'new servergroup'(id:2) was deleted by 'admin'(id:2)"),
+        serverGroupParser.parseMessageServerGroupDeletion("servergroup 'new servergroup'(id:2) was deleted by 'admin'(id:2)"),
         {
           ServerGroupID: 2,
           ServerGroupName: 'new servergroup'
@@ -32,10 +32,10 @@ describe('serverGroupParser', function () {
     });
   });
 
-  describe('parseServerGroupRenaming', function () {
+  describe('parseMessageServerGroupRenaming', function () {
     it('should return the specified object', function () {
       assert.deepEqual(
-        serverGroupParser.parseServerGroupRenaming("servergroup 'new servergroup'(id:2) was renamed to 'old servergroup' by 'admin'(id:2)"),
+        serverGroupParser.parseMessageServerGroupRenaming("servergroup 'new servergroup'(id:2) was renamed to 'old servergroup' by 'admin'(id:2)"),
         {
           ServerGroupID: 2,
           ServerGroupName: 'old servergroup'
@@ -43,10 +43,10 @@ describe('serverGroupParser', function () {
     });
   });
 
-  describe('parseServerGroupCopying', function () {
+  describe('parseMessageServerGroupCopying', function () {
     it('should return the specified object', function () {
       assert.deepEqual(
-        serverGroupParser.parseServerGroupCopying("servergroup 'old servergroup'(id:2) was copied by 'admin'(id:2) to 'copy of old servergroup'(id:3)"),
+        serverGroupParser.parseMessageServerGroupCopying("servergroup 'old servergroup'(id:2) was copied by 'admin'(id:2) to 'copy of old servergroup'(id:3)"),
         {
           ServerGroupID: 3,
           ServerGroupName: 'copy of old servergroup'